backend/requirements.txt
# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2

# Vector Database
weaviate-client==3.24.1

# Document Processing
langchain==0.0.325
python-dotenv==1.0.0
pypdf==3.17.0
docx2txt==0.8
pytesseract==0.3.10
pillow==10.1.0
PyMuPDF==1.23.7  # For PDF processing
pandas==2.1.3  # For table handling

# NLP and ML
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
spacy==3.7.2
scikit-learn==1.3.2

# Utilities
python-jose[cryptography]==3.3.0  # For JWT handling
passlib[bcrypt]==1.7.4  # For password hashing
tenacity==8.2.3  # For retrying operations
python-magic==0.4.27  # For file type detection
aiofiles==23.2.1  # For async file operations
